short

+++ lcs-serial.c
@@ -22,7 +22,7 @@ int main(int argc, char *argv[]){
 	int size_x, size_y, i, j;
 	int size_xx, size_yy;
 	char *x=NULL, *y=NULL, *z=NULL; 
-	int **c = NULL;
+	short **c = NULL;
 	
 	
 	if (argc !=2)
@@ -80,7 +80,7 @@ int main(int argc, char *argv[]){
 	fgets(buffer, size_yy,  f);
 	sscanf(buffer, "%s\n", y);
 
-	c = (int **)calloc((size_x+2), sizeof(int*));
+	c = (short **)calloc((size_x+2), sizeof(short*));
 	if (c == NULL)
 	{
 		fprintf(stdout, "Error in c malloc\n");
@@ -89,7 +89,7 @@ int main(int argc, char *argv[]){
 	
 	for(i = 0; i <= size_x; i++)
 	{
-		c[i] = (int *)calloc((size_y+2), sizeof(int));
+		c[i] = (short *)calloc((size_y+2), sizeof(short));
 		if (c[i] == NULL)
 		{
 		fprintf(stdout, "Error in c[i] calloc\n");
